rename cart_copy to fileCentricCart
