+++ src/pages/fileCentricCart/assets/arrowUp.png
